Fix a bug with wrong calculation for the amount of particles
I was casting a double to an int, but the double was to big to be stored by an int.

+++ app/src/main/java/com/teinproductions/tein/molu/Element.java
@@ -140,8 +140,8 @@ public enum Element {
         return givenMol * this.mass;
     }
 
-    public Integer calculateParticlesWhenMolGiven(Double givenMol){
-        return (int) (givenMol * nA);
+    public Double calculateParticlesWhenMolGiven(Double givenMol){
+        return givenMol * nA;
     }
 
     public Double calculateMolWhenGramGiven(Double givenGram){
